+++ tests/Console/ModelsCommand/Attributes/__snapshots__/Test__test__1.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Attributes\Models;
+
+use Illuminate\Database\Eloquent\Casts\Attribute;
+use Illuminate\Database\Eloquent\Model;
+
+/**
+ * Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Attributes\Models\Simple
+ *
+ * @property integer $id
+ * @property string|null $name
+ * @method static \Illuminate\Database\Eloquent\Builder|Simple newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|Simple newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|Simple query()
+ * @method static \Illuminate\Database\Eloquent\Builder|Simple whereId($value)
+ * @mixin \Eloquent
+ */
+class Simple extends Model
+{
+    public function name(): Attribute
+    {
+        return new Attribute(
+            function (?string $name): ?string {
+                return $name;
+            },
+            function (?string $name): ?string {
+                return $name === null ? null : ucfirst($name);
+            }
+        );
+    }
+}